@@ -9,21 +9,24 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@ant-design/react-slick": "^1.0.2",
     "@next-auth/mongodb-adapter": "^1.1.3",
     "@next-auth/prisma-adapter": "^1.0.7",
-    "@prisma/client": "^5.2.0",
+    "@prisma/client": "^5.4.2",
+    "@radix-ui/react-dialog": "^1.0.5",
     "@radix-ui/react-dropdown-menu": "^2.0.5",
+    "@radix-ui/react-label": "^2.0.2",
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-toast": "^1.1.4",
     "@stripe/stripe-js": "^1.54.1",
     "@uidotdev/usehooks": "^2.0.1",
     "@uploadthing/react": "^5.2.0",
-    "axios": "^1.4.0",
     "babel-plugin-styled-components": "^2.1.3",
+    "bcrypt": "^5.1.1",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.0.0",
     "eslint": "8.39.0",
-    "eslint-config-next": "13.4.6",
+    "eslint-config-next": "14.0.1",
     "framer-motion": "^10.12.17",
     "lucide-react": "^0.263.0",
     "micro": "^10.0.1",
@@ -31,17 +34,20 @@
     "mongodb": "^5.3.0",
     "mongoose": "^7.3.3",
     "multiparty": "^4.2.3",
-    "next": "13.4.7",
+    "next": "14.0.1",
     "next-auth": "^4.22.1",
     "next-seo": "^6.1.0",
+    "pnpm": "^8.9.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-dropzone": "^14.2.3",
     "react-icons": "^4.10.1",
+    "react-password-toggle-icon": "^1.1.5",
     "react-slick": "^0.29.0",
     "react-sortablejs": "^6.1.4",
     "react-spinners": "^0.13.8",
     "react-sweetalert2": "^0.5.2",
+    "redaxios": "^0.5.1",
     "rsuite": "^5.35.1",
     "slick-carousel": "^1.8.1",
     "stripe": "^12.14.0",
@@ -54,7 +60,7 @@
   "devDependencies": {
     "autoprefixer": "^10.4.14",
     "postcss": "^8.4.25",
-    "prisma": "5.2.0",
+    "prisma": "5.4.2",
     "prisma-client": "^0.0.0",
     "tailwindcss": "^3.3.2",
     "webpack-filter-warnings-plugin": "^1.2.1"
