@@ -1,25 +1,23 @@
-import Hero from "../../components/front/Hero";
-import {ProductsHome} from "@/src/components/front/ProductsHome";
-import {getServerSession} from "next-auth/next";
-import {authOptions} from "@/src/lib/auth";
-import { redirect } from "next/dist/server/api-utils";
+import Hero from "../../components/front/Hero"
+import { ProductsHome } from "@/src/components/front/ProductsHome"
+import { getServerSession } from "next-auth/next"
+import { redirect } from "next/navigation"
 // import {NextSeo} from "next-seo";
-// import {redirect} from "next/navigation";
+import { cookies } from "next/headers"
 
 export default async function Homepage() {
-    const session = await getServerSession(authOptions)
+  // const cookieSession = cookies().get("session")
 
-    if (!session) {
-        return redirect('/signin')
-    }
+  // if (!cookieSession) {
+  //   redirect("/login")
+  // }
 
-    return (
-        <>
-            <Hero/>
-            <div className="bg-slate-950">
-                <ProductsHome/>
-            </div>
-        </>
-    )
+  return (
+    <>
+      <Hero />
+      <div className="bg-slate-950">
+        <ProductsHome />
+      </div>
+    </>
+  )
 }
-
