@@ -1,7 +1,6 @@
-import "../../styles/globals.css";
-import "slick-carousel/slick/slick.css";
-import "slick-carousel/slick/slick-theme.css";
-import Provider from "./provider";
+import "../../styles/globals.css"
+import { Toaster } from "@/src/components/ui/toaster"
+import Provider from "./provider"
 // import Head from "next/head";
 // import Head from "next/head";
 // import {Metadata} from "next";
@@ -11,19 +10,20 @@ import Provider from "./provider";
 // import {NextSeo} from "next-seo";
 
 export const metadata = {
-    title: 'Site E-commerce',
-    descripton: 'Site E-commerce'
+  title: "Site E-commerce",
+  descripton: "Site E-commerce",
 }
 
-export default function RootLayout({children}) {
-    //
-    return (
-        <html lang="fr" data-theme="corporate">
-            <Provider>
-                <body>
-                    {children}
-                </body>
-            </Provider>
-        </html>
-    )
+export default function RootLayout({ children }) {
+  //
+  return (
+    <html lang="fr" data-theme="corporate">
+      <Provider>
+        <body>
+          <Toaster />
+          {children}
+        </body>
+      </Provider>
+    </html>
+  )
 }
