@@ -1,11 +1,12 @@
-'use client'
+"use client"
+import Spinner from "@/src/components/Spinner"
 
 export default function Loading() {
-    return (
-        <>
-            <div className="h-screen w-screen flex justify-center items-center">
-                <span className="loading loading-dots loading-md"></span>
-            </div>
-        </>
-    )
+  return (
+    <>
+      <div className="h-screen w-full flex justify-center items-center">
+        <Spinner />
+      </div>
+    </>
+  )
 }
