@@ -1,50 +1,70 @@
-'use client'
-import {CartContext} from "@/src/components/CartContext";
-import {useContext} from "react";
+"use client"
+import { CartContext } from "@/src/components/CartContext"
+import { useContext } from "react"
 import { Button } from "@/src/components/ui/button"
 import { useToast } from "@/src/components/ui/use-toast"
-import { Toaster } from "@/src/components/ui/toaster"
 
-export const BtnAddToCartPrimary = ({featureId,featureName}) => {
-    const { toast } = useToast()
-    const {addProduct} = useContext(CartContext)
-    return (
-        <Button onClick={() =>{
-            addProduct(featureId)
-            toast({
-                title: "Success",
-                description: `Vous avez ajouté ${featureName} au panier`,
-              })
-            }}>
-            <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5}
-                 stroke="currentColor" className="w-6 h-6">
-                <path strokeLinecap="round" strokeLinejoin="round"
-                      d="M2.25 3h1.386c.51 0 .955.343 1.087.835l.383 1.437M7.5 14.25a3 3 0 00-3 3h15.75m-12.75-3h11.218c1.121-2.3 2.1-4.684 2.924-7.138a60.114 60.114 0 00-16.536-1.84M7.5 14.25L5.106 5.272M6 20.25a.75.75 0 11-1.5 0 .75.75 0 011.5 0zm12.75 0a.75.75 0 11-1.5 0 .75.75 0 011.5 0z"/>
-            </svg>
-            Add to cart
-        </Button>
-    );
-};
+export const BtnAddToCartPrimary = ({ featureId, featureName }) => {
+  const { toast } = useToast()
+  const { addProduct } = useContext(CartContext)
+  return (
+    <Button
+      onClick={() => {
+        addProduct(featureId)
+        toast({
+          title: "Success",
+          description: `Vous avez ajouté ${featureName} au panier`,
+        })
+      }}
+    >
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        fill="none"
+        viewBox="0 0 24 24"
+        strokeWidth={1.5}
+        stroke="currentColor"
+        className="w-6 h-6"
+      >
+        <path
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          d="M2.25 3h1.386c.51 0 .955.343 1.087.835l.383 1.437M7.5 14.25a3 3 0 00-3 3h15.75m-12.75-3h11.218c1.121-2.3 2.1-4.684 2.924-7.138a60.114 60.114 0 00-16.536-1.84M7.5 14.25L5.106 5.272M6 20.25a.75.75 0 11-1.5 0 .75.75 0 011.5 0zm12.75 0a.75.75 0 11-1.5 0 .75.75 0 011.5 0z"
+        />
+      </svg>
+      Add to cart
+    </Button>
+  )
+}
 
-export const BtnAddToCartSecondary = ({id, title}) => {
-    const { toast } = useToast()
-    const {addProduct} = useContext(CartContext)
-    return (
-        <Button onClick={() => {
-                                addProduct(id)
-                                toast({
-                                    title: "Success",
-                                    description: `Vous avez ajouté ${title} au panier`,
-                                })
-                                }}
-                className="btn btn-outline flex-nowrap gap-2"
-                >
-            <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5}
-                 stroke="currentColor" className="w-6 h-6">
-                <path strokeLinecap="round" strokeLinejoin="round"
-                      d="M2.25 3h1.386c.51 0 .955.343 1.087.835l.383 1.437M7.5 14.25a3 3 0 00-3 3h15.75m-12.75-3h11.218c1.121-2.3 2.1-4.684 2.924-7.138a60.114 60.114 0 00-16.536-1.84M7.5 14.25L5.106 5.272M6 20.25a.75.75 0 11-1.5 0 .75.75 0 011.5 0zm12.75 0a.75.75 0 11-1.5 0 .75.75 0 011.5 0z"/>
-            </svg>
-            <span className="flex flex-nowrap whitespace-nowrap">Add to cart</span>
-        </Button>
-    );
-};
+export const BtnAddToCartSecondary = ({ id, title }) => {
+  const { toast } = useToast()
+  const { addProduct } = useContext(CartContext)
+  return (
+    <Button
+      onClick={() => {
+        addProduct(id)
+        toast({
+          title: "Success",
+          description: `Vous avez ajouté ${title} au panier`,
+        })
+      }}
+      className="btn btn-outline flex-nowrap gap-2"
+    >
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        fill="none"
+        viewBox="0 0 24 24"
+        strokeWidth={1.5}
+        stroke="currentColor"
+        className="w-6 h-6"
+      >
+        <path
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          d="M2.25 3h1.386c.51 0 .955.343 1.087.835l.383 1.437M7.5 14.25a3 3 0 00-3 3h15.75m-12.75-3h11.218c1.121-2.3 2.1-4.684 2.924-7.138a60.114 60.114 0 00-16.536-1.84M7.5 14.25L5.106 5.272M6 20.25a.75.75 0 11-1.5 0 .75.75 0 011.5 0zm12.75 0a.75.75 0 11-1.5 0 .75.75 0 011.5 0z"
+        />
+      </svg>
+      <span className="flex flex-nowrap whitespace-nowrap">Add to cart</span>
+    </Button>
+  )
+}
